Changing the classes desgin based on team conversation
