+++ src/rocjpeg/roc_pyjpeg_codestream.cpp
@@ -20,7 +20,6 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 THE SOFTWARE.
 */
 
-#include "roc_pyjpeg.h"
 #include "roc_pyjpeg_decoder.h"
 #include "common/roc_pybuffer.h"
 #include "roc_pyjpeg_codestream.h"
@@ -67,41 +66,7 @@ void CodeStream::exportToPython(py::module& m) {
 
             Args:
                 filename: The file path to the encoded stream data.
-            )pbdoc")
-        .def_property_readonly("height", &CodeStream::height, 
-            R"pbdoc(
-            The vertical dimension of the entire image in pixels.
-            )pbdoc")
-        .def_property_readonly("width", &CodeStream::width, 
-            R"pbdoc(
-            The horizontal dimension of the entire image in pixels.
-            )pbdoc")
-            ;
-}
-
-CodeStream& CodeStream::operator=(const CodeStream& other) {
-    if (this == &other)
-        return *this; // handle self-assignment
-    // decoder members
-    m_width = other.m_width;
-    m_height = other.m_height;
-    valid = other.valid;
-    num_channels = other.num_channels;
-    subsampling = other.subsampling;
-    // w/h arrays
-    std::copy(std::begin(other.widths), std::end(other.widths), std::begin(widths));
-    std::copy(std::begin(other.heights), std::end(other.heights), std::begin(heights));
-    std::copy(std::begin(other.channel_sizes), std::end(other.channel_sizes), std::begin(channel_sizes));
-    // Shared data buffer (shared_ptr)
-    file_data = other.file_data;
-    // Pybind11 refs
-    data_ref_bytes_ = other.data_ref_bytes_;
-    data_ref_arr_ = other.data_ref_arr_;
-    // GPU handles - decoder
-    stream_handle = other.stream_handle;
-    decode_params = other.decode_params;
-    output_image = other.output_image;
-    return *this;
+            )pbdoc");
 }
 
 int CodeStream::ReadFromFile(const std::filesystem::path& filename, std::shared_ptr<std::vector<char>>& file_data, int& file_size) {
@@ -128,11 +93,6 @@ int CodeStream::ReadFromFile(const std::filesystem::path& filename, std::shared_
 
 // Use the dat and its size if valid, otherwise use the file to load the data
 int CodeStream::InitializeSingleImage(const std::filesystem::path& filename, const unsigned char* data, int data_size) {
-    // default, reset
-    memset(&decode_params, 0, sizeof(RocJpegDecodeParams));
-    memset(&output_image, 0, sizeof(RocJpegImage));
-    decode_params.output_format = Decoder::GetFormat();
-    num_channels = 0;
     // File sanity check
     if(!filename.empty()) {
         if(!std::filesystem::exists(filename)) {
@@ -149,11 +109,6 @@ int CodeStream::InitializeSingleImage(const std::filesystem::path& filename, con
             return ret;
         }
     }
-    // Create Stream - release/de-alloc if previously used
-    if(stream_handle != nullptr) {
-        PY_CHECK_ROCJPEG(rocJpegStreamDestroy(stream_handle));
-        stream_handle = nullptr;
-    }
     RocJpegStatus rocjpeg_status = ROCJPEG_STATUS_NOT_INITIALIZED;
     rocjpeg_status = rocJpegStreamCreate(&stream_handle);
     if (rocjpeg_status != ROCJPEG_STATUS_SUCCESS) {
@@ -166,46 +121,6 @@ int CodeStream::InitializeSingleImage(const std::filesystem::path& filename, con
         std::cerr << "ERROR: Failed to parse the input jpeg stream with " << rocJpegGetErrorName(rocjpeg_status) << "Input File : " << (!filename.empty() ? filename : "") << std::endl;
         return Release();
     }
-    // Get Image Info
-    uint8_t num_components = 0;
-    subsampling = ROCJPEG_CSS_UNKNOWN;
-    RocJpegHandle jpeg_handle = Decoder::GetHandle();
-    rocjpeg_status = rocJpegGetImageInfo(jpeg_handle, stream_handle, &num_components, &subsampling, widths, heights);
-    if (rocjpeg_status != ROCJPEG_STATUS_SUCCESS) {
-        std::cerr << "ERROR: Failed to  get image info with " << rocJpegGetErrorName(rocjpeg_status) << std::endl;
-        return Release();
-    }
-    // Check limits of w/h & subsampling
-    PyRocJpegUtils rocjpeg_utils;
-    if (widths[0] < 64 || heights[0] < 64) {
-        std::cerr << "The image resolution is not supported by VCN Hardware: " << (!filename.empty() ? filename : "") << std::endl;
-        return Release();
-    }
-    if (subsampling == ROCJPEG_CSS_411 || subsampling == ROCJPEG_CSS_UNKNOWN) {
-        std::cerr << "The image resolution is not supported by VCN Hardware: " << (!filename.empty() ? filename : "") << std::endl;
-        return Release();
-    }    
-    // save the output w/h to the inner store
-    m_width = widths[0];
-    m_height = heights[0];
-    // Get Channel Pitch And Sizes
-    if (rocjpeg_utils.GetChannelPitchAndSizes(decode_params, subsampling, widths, heights, num_channels, output_image, channel_sizes)) {
-        std::cerr << "ERROR: Failed to get the channel pitch and sizes" << std::endl;
-        return Release();
-    }
-    // allocate memory for each channel
-    hipError_t hip_status = hipSuccess;
-    for (int i = 0; i < num_channels; i++) {
-        if (output_image.channel[i] != nullptr) {
-            hip_status = hipFree((void *)output_image.channel[i]);
-            if (hip_status != hipSuccess)
-                return Release();
-            output_image.channel[i] = nullptr;
-        }
-        hip_status = hipMalloc(&output_image.channel[i], channel_sizes[i]);
-        if (hip_status != hipSuccess)
-            return Release();
-    }
     return EXIT_SUCCESS;
 }
 
@@ -214,12 +129,6 @@ int CodeStream::Release() {
     if(stream_handle) {
         RocJpegStatus rocjpeg_status = rocJpegStreamDestroy(stream_handle);
         stream_handle = nullptr;
-    }
-    for (int i = 0; i < num_channels; i++) {
-        if (output_image.channel[i] != nullptr) {
-            hip_status = hipFree((void *)output_image.channel[i]);
-            output_image.channel[i] = nullptr;
-        }
     }
     return EXIT_FAILURE;
 }
@@ -264,12 +173,5 @@ CodeStream::CodeStream() {
 }
 
 CodeStream::~CodeStream() {
-}
-
-int CodeStream::width() {
-    return m_width;
-}
-
-int CodeStream::height() {
-    return m_height;
+    Release();
 }
