+++ src/rocjpeg/roc_pyjpeg_codestream.h
@@ -26,8 +26,8 @@ THE SOFTWARE.
 #pragma once
 
 #include <filesystem>
-#include "roc_pyjpeg.h"
 #include "common/roc_pybuffer.h"
+#include "rocjpeg/rocjpeg.h"
 
 class CodeStream {
 
@@ -38,27 +38,17 @@ public:
     CodeStream(py::array_t<uint8_t>);
     ~CodeStream();
     CodeStream();
-    CodeStream& operator=(const CodeStream& other);
+
     static void exportToPython(py::module& m);
 
-    int width();
-    int height();
-    
     // related descriptor of 'this' image
     RocJpegStreamHandle stream_handle = nullptr;
-    RocJpegDecodeParams decode_params;
-    RocJpegImage output_image;
-    RocJpegChromaSubsampling subsampling;
     std::shared_ptr<std::vector<char>> file_data;
-    int file_size;
-    uint32_t widths[ROCJPEG_MAX_COMPONENT] = {};
-    uint32_t heights[ROCJPEG_MAX_COMPONENT] = {};    
-    uint32_t channel_sizes[ROCJPEG_MAX_COMPONENT] = {};
-    uint32_t num_channels = 0;
+    
     void SetValid(bool state) {valid = state;};
-    bool IsValid() {return valid;};
+    const bool IsValid() const {return valid;};
 
-    // flag = true indicates this instance of the code_stream is valid to use
+    // flag = true code_stream is valid to use
     // when = false means the file/data associated with it found invalid
 protected:
     bool valid;
@@ -66,8 +56,6 @@ protected:
 private:
     py::bytes data_ref_bytes_;
     py::array_t<uint8_t> data_ref_arr_;
-    int m_width;
-    int m_height;
     int Release();
     int ReadFromFile(const std::filesystem::path& filename, std::shared_ptr<std::vector<char>>& file_data, int& file_size);
     int InitializeSingleImage(const std::filesystem::path& filename, const unsigned char* data, int data_size);
